Algoritmos/auralis/requirements.txt
paho-mqtt==1.6.1
celery==5.2.7
redis==4.5.1
SQLAlchemy==1.4.46
mysqlclient==2.1.1 